+++ consts.js
@@ -1,0 +1,4 @@
+module.exports = {
+    gamemodes: ['ffa', 'survival', 'teams', '4teams', 'dom', 'tag', 'maze', 'sandbox'],
+    regions: ['do-sfo','do-nyc','do-fra','do-sgp']
+};